Python testing: move PICS functions
We're going to need these for some new tooling, so moving them
into their own file to avoid needing to pull in the entirety
of matter_testing_support

Only changes are moves and function renames.

test: TC_pics_checker.py
